# PERF: groupby.quantile