--- a/pandas/core/groupby/groupby.py
+++ b/pandas/core/groupby/groupby.py
@@ -50,6 +50,7 @@ from pandas._typing import (
     RandomState,
     Scalar,
     T,
+    npt,
 )
 from pandas.compat.numpy import function as nv
 from pandas.errors import AbstractMethodError
@@ -59,6 +60,7 @@ from pandas.util._decorators import (
     cache_readonly,
     doc,
 )
+from pandas.util._exceptions import find_stack_level
 
 from pandas.core.dtypes.common import (
     is_bool_dtype,
@@ -1101,7 +1103,9 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
         raise AbstractMethodError(self)
 
     def _wrap_aggregated_output(
-        self, output: Series | DataFrame | Mapping[base.OutputKey, ArrayLike]
+        self,
+        output: Series | DataFrame | Mapping[base.OutputKey, ArrayLike],
+        qs: npt.NDArray[np.float64] | None = None,
     ):
         """
         Wraps the output of GroupBy aggregations into the expected result.
@@ -1130,8 +1134,17 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
             #   enforced in __init__
             self._insert_inaxis_grouper_inplace(result)
             result = result._consolidate()
+            index = Index(range(self.grouper.ngroups))
+
         else:
-            result.index = self.grouper.result_index
+            index = self.grouper.result_index
+
+        if qs is not None:
+            # We get here with len(qs) != 1 and not self.as_index
+            #  in test_pass_args_kwargs
+            index = _insert_quantile_level(index, qs)
+
+        result.index = index
 
         if self.axis == 1:
             # Only relevant for DataFrameGroupBy, no-op for SeriesGroupBy
@@ -1141,7 +1154,7 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
                 result.index = self.obj.index.copy()
                 # TODO: Do this more systematically
 
-        return self._reindex_output(result)
+        return self._reindex_output(result, qs=qs)
 
     def _wrap_transformed_output(self, output: Mapping[base.OutputKey, ArrayLike]):
         raise AbstractMethodError(self)
@@ -2480,7 +2493,6 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
         a    2.0
         b    3.0
         """
-        from pandas import concat
 
         def pre_processor(vals: ArrayLike) -> tuple[np.ndarray, np.dtype | None]:
             if is_object_dtype(vals):
@@ -2511,7 +2523,7 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
 
             return out, inference
 
-        def post_processor(vals: np.ndarray, inference: type | None) -> np.ndarray:
+        def post_processor(vals: np.ndarray, inference: np.dtype | None) -> np.ndarray:
             if inference:
                 # Check for edge case
                 if not (
@@ -2523,63 +2535,71 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
             return vals
 
         if is_scalar(q):
-            return self._get_cythonized_result(
-                libgroupby.group_quantile,
-                aggregate=True,
-                numeric_only=False,
-                needs_values=True,
-                needs_mask=True,
-                cython_dtype=np.dtype(np.float64),
-                pre_processing=pre_processor,
-                post_processing=post_processor,
-                q=q,
-                interpolation=interpolation,
-            )
-        else:
-            results = [
-                self._get_cythonized_result(
-                    libgroupby.group_quantile,
-                    aggregate=True,
-                    needs_values=True,
-                    needs_mask=True,
-                    cython_dtype=np.dtype(np.float64),
-                    pre_processing=pre_processor,
-                    post_processing=post_processor,
-                    q=qi,
-                    interpolation=interpolation,
-                )
-                for qi in q
-            ]
-            result = concat(results, axis=self.axis, keys=q)
-            # fix levels to place quantiles on the inside
-            # TODO(GH-10710): Ideally, we could write this as
-            #  >>> result.stack(0).loc[pd.IndexSlice[:, ..., q], :]
-            #  but this hits https://github.com/pandas-dev/pandas/issues/10710
-            #  which doesn't reorder the list-like `q` on the inner level.
-            order = list(range(1, result.axes[self.axis].nlevels)) + [0]
+            res = self.quantile([q], interpolation=interpolation)
+            nlevels = res.index.nlevels
+            return res.droplevel(nlevels - 1, axis=0)
 
-            # temporarily saves the index names
-            index_names = np.array(result.axes[self.axis].names)
+        qs = np.array(q, dtype=np.float64)
+        ids, _, ngroups = self.grouper.group_info
+        nqs = len(qs)
 
-            # set index names to positions to avoid confusion
-            result.axes[self.axis].names = np.arange(len(index_names))
+        func = partial(
+            libgroupby.group_quantile, labels=ids, qs=qs, interpolation=interpolation
+        )
 
-            # place quantiles on the inside
-            if isinstance(result, Series):
-                result = result.reorder_levels(order)
+        def blk_func(values: ArrayLike) -> ArrayLike:
+            mask = isna(values)
+            vals, inference = pre_processor(values)
+
+            ncols = 1
+            if vals.ndim == 2:
+                ncols = vals.shape[0]
+
+            out = np.empty((ncols, ngroups, nqs), dtype=np.float64)
+
+            if vals.ndim == 1:
+                func(out[0], values=vals, mask=mask)
             else:
-                result = result.reorder_levels(order, axis=self.axis)
+                for i in range(ncols):
+                    func(out[i], values=vals[i], mask=mask[i])
 
-            # restore the index names in order
-            result.axes[self.axis].names = index_names[order]
+            if vals.ndim == 1:
+                out = out[0].ravel("K")
+            else:
+                out = out.reshape(ncols, ngroups * nqs)
+            return post_processor(out, inference)
 
-            # reorder rows to keep things sorted
-            indices = (
-                np.arange(result.shape[self.axis])
-                .reshape([len(q), self.ngroups])
-                .T.flatten()
+        obj = self._obj_with_exclusions
+        is_ser = obj.ndim == 1
+        if is_ser:
+            # i.e. SeriesGroupBy
+            mgr = obj.to_frame()._mgr
+        else:
+            mgr = self._get_data_to_aggregate()
+
+        res_mgr = mgr.grouped_reduce(blk_func, ignore_failures=True)
+        if len(res_mgr.items) != len(mgr.items):
+            warnings.warn(
+                "Dropping invalid columns in "
+                f"{type(self).__name__}.quantile is deprecated. "
+                "In a future version, a TypeError will be raised. "
+                "Before calling .quantile, select only columns which "
+                "should be valid for the function.",
+                FutureWarning,
+                stacklevel=find_stack_level(),
             )
-            return result.take(indices, axis=self.axis)
+            if len(res_mgr.items) == 0:
+                # re-call grouped_reduce to get the desired exception message
+                mgr.grouped_reduce(blk_func, ignore_failures=False)
+
+        if is_ser:
+            res = obj._constructor_expanddim(res_mgr)
+            res = res[res.columns[0]]  # aka res.squeeze()
+            res.name = obj.name
+        else:
+            res = obj._constructor(res_mgr)
+
+        return self._wrap_aggregated_output(res, qs=qs)
 
     @final
     @Substitution(name="groupby")
@@ -2994,6 +3014,7 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
                     result = result.reshape((result_sz, ncols))
                 else:
                     result = result.reshape(-1, 1)
+
             func = partial(base_func, out=result)
 
             inferences = None
@@ -3088,7 +3109,7 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
                     f"Before calling .{howstr}, select only columns which "
                     "should be valid for the function.",
                     FutureWarning,
-                    stacklevel=3,
+                    stacklevel=find_stack_level(),
                 )
                 continue
 
@@ -3258,7 +3279,10 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
 
     @final
     def _reindex_output(
-        self, output: OutputFrameOrSeries, fill_value: Scalar = np.NaN
+        self,
+        output: OutputFrameOrSeries,
+        fill_value: Scalar = np.NaN,
+        qs: npt.NDArray[np.float64] | None = None,
     ) -> OutputFrameOrSeries:
         """
         If we have categorical groupers, then we might want to make sure that
@@ -3277,6 +3301,8 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
             Object resulting from grouping and applying an operation.
         fill_value : scalar, default np.NaN
             Value to use for unobserved categories if self.observed is False.
+        qs : np.ndarray[float64] or None, default None
+            quantile values, only relevant for quantile.
 
         Returns
         -------
@@ -3300,9 +3326,11 @@ class GroupBy(BaseGroupBy[FrameOrSeries]):
             return output
 
         levels_list = [ping.group_index for ping in groupings]
-        index, _ = MultiIndex.from_product(
-            levels_list, names=self.grouper.names
-        ).sortlevel()
+        names = self.grouper.names
+        if qs is not None:
+            levels_list.append(qs)
+            names = names + [None]
+        index, _ = MultiIndex.from_product(levels_list, names=names).sortlevel()
 
         if self.as_index:
             d = {
@@ -3517,3 +3545,22 @@ def get_groupby(
         mutated=mutated,
         dropna=dropna,
     )
+
+
+def _insert_quantile_level(idx: Index, qs: npt.NDArray[np.float64]) -> MultiIndex:
+    """
+    Insert the sequence 'qs' of quantiles as the inner-most level of a MultiIndex.
+
+    The quantile level in the MultiIndex is a repeated copy of 'qs'.
+    """
+    nqs = len(qs)
+
+    if idx._is_multi:
+        idx = cast(MultiIndex, idx)
+        lev_codes, lev = Index(qs).factorize()
+        levels = list(idx.levels) + [lev]
+        codes = [np.repeat(x, nqs) for x in idx.codes] + [np.tile(lev_codes, len(idx))]
+        mi = MultiIndex(levels=levels, codes=codes, names=idx.names + [None])
+    else:
+        mi = MultiIndex.from_product([idx, qs])
+    return mi