@jreback 2021-09-09
prob worth a whatsnew note as large change in perf